已在两个项目（windows）应用
*添加串口（初步）

+++ base/functorloop.h
@@ -44,10 +44,10 @@ public:
 	{ return queue_.run(); }
 
 	int getPending()
-	{ return queue_.size(); }
+	{ return (int)queue_.size(); }
 
 	int getActPending()
-	{ return queue_.actSize(); }
+	{ return (int)queue_.actSize(); }
 
 private:
 	void defaultLoop();
